Reorganize app windows and restore UI window utilities
