+++ static/app/js/windows/llm_service_admin.js
